requirements.txt
streamlit==1.29.0
yfinance==0.2.33
pandas==2.1.4
plotly==5.18.0
python-dateutil==2.8.2
requests==2.31.0
brapi==2.0.3
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0